update base64ToString method
