+++ src/utils/jwt-utils.ts
@@ -1,3 +1,5 @@
+import { base64toString } from './string-utils.js';
+
 /**
  * Laserfiche Cloud endpoints
  */
@@ -81,27 +83,6 @@ export function getLfEndpoints(accountId: string, devEnvironmentSubDomain?: stri
     repositoryApiBaseUrl: `https://api.${regionSpecificHostName}/repository/`,
     wsignoutUrl: `https://accounts.${regionSpecificHostName}/WebSTS/?wa=wsignout1.0`,
   };
-}
-
-/**
- * Decodes a string of data encoded using Base64 encoding
- * @param base64String
- * @returns
- * @example
- * ```typescript
- * base64toString('dGVzdA=='); // => 'test';
- * ```
- */
-export function base64toString(base64String: string): string {
-  if (Buffer) {
-    return Buffer.from(base64String, 'base64').toString();
-  }
-  else if (window) {
-    return window.atob(base64String);
-  }
-  else {
-    throw new Error('No method exists to decode base64');
-  }
 }
 
 /**
